# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Regolith API",
    "description": "Regolith food catalog API",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/profiles/{profileId}": {
      "put": {
        "operationId": "ensureProfile",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            },
            "description": "Profile is ready"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid profile identifier"
          }
        },
        "tags": [
          "profiles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ]
      }
    },
    "/v1/profiles/{profileId}/nutrition-plan": {
      "get": {
        "operationId": "getNutritionPlan",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionPlanResponse"
                }
              }
            },
            "description": "Saved nutrition plan, or null before onboarding is complete"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid profile identifier"
          }
        },
        "tags": [
          "profiles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ]
      },
      "put": {
        "operationId": "saveNutritionPlan",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionPlan"
                }
              }
            },
            "description": "Calculated and saved nutrition plan"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid or unsupported nutrition plan"
          }
        },
        "tags": [
          "profiles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveNutritionPlan"
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/food-log": {
      "get": {
        "operationId": "listFoodLog",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodLogResponse"
                }
              }
            },
            "description": "Food log entries in the requested time range"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid profile or time range"
          }
        },
        "tags": [
          "food-log"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": true
          }
        ]
      },
      "post": {
        "operationId": "createFoodLogEntry",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodLogEntry"
                }
              }
            },
            "description": "Created or updated food log entry"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid food log entry"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Catalog food not found"
          }
        },
        "tags": [
          "food-log"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFoodLogEntry"
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/food-log/{entryId}": {
      "delete": {
        "operationId": "deleteFoodLogEntry",
        "responses": {
          "204": {
            "description": "Food log entry deleted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid identifiers"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Food log entry not found"
          }
        },
        "tags": [
          "food-log"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entryId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ]
      }
    },
    "/v1/profiles/{profileId}/weight-log": {
      "get": {
        "operationId": "listWeightLog",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightLogResponse"
                }
              }
            },
            "description": "Weight entries in the requested time range"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid profile or time range"
          }
        },
        "tags": [
          "weight-log"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": true
          }
        ]
      },
      "post": {
        "operationId": "saveWeightLogEntry",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightLogEntry"
                }
              }
            },
            "description": "Created or updated weight entry"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid weight entry"
          }
        },
        "tags": [
          "weight-log"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeightLogEntry"
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/weight-log/{entryId}": {
      "delete": {
        "operationId": "deleteWeightLogEntry",
        "responses": {
          "204": {
            "description": "Weight entry deleted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid identifiers"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Weight entry not found"
          }
        },
        "tags": [
          "weight-log"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entryId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ]
      }
    },
    "/v1/profiles/{profileId}/workouts": {
      "get": {
        "operationId": "listWorkouts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            },
            "description": "Workout sessions in the requested time range"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid profile or time range"
          }
        },
        "tags": [
          "workouts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": true
          }
        ]
      }
    },
    "/v1/profiles/{profileId}/workouts/{sessionId}": {
      "put": {
        "operationId": "saveWorkout",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            },
            "description": "Saved workout session"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid workout"
          }
        },
        "tags": [
          "workouts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "sessionId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveWorkout"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteWorkout",
        "responses": {
          "204": {
            "description": "Workout deleted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid identifiers"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Workout not found"
          }
        },
        "tags": [
          "workouts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "sessionId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Profile": {
        "type": "object",
        "properties": {
          "profileId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "profileId"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "NutritionPlan": {
        "type": "object",
        "properties": {
          "birthDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "currentWeightKg": {
            "type": "number",
            "minimum": 30,
            "maximum": 350
          },
          "dailyActivity": {
            "enum": [
              "mostly_sedentary",
              "moderately_active",
              "very_active"
            ],
            "type": "string"
          },
          "exerciseFrequency": {
            "enum": [
              "none",
              "one_to_three",
              "four_to_six",
              "seven_plus"
            ],
            "type": "string"
          },
          "heightCm": {
            "type": "number",
            "minimum": 100,
            "maximum": 250
          },
          "metabolicSex": {
            "enum": [
              "female",
              "male"
            ],
            "type": "string"
          },
          "targetWeightKg": {
            "type": "number",
            "minimum": 30,
            "maximum": 350
          },
          "weeklyWeightChangePercent": {
            "type": "number",
            "minimum": 0,
            "maximum": 1.25
          },
          "weightGoal": {
            "enum": [
              "lose",
              "maintain",
              "gain"
            ],
            "type": "string"
          },
          "calculatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "calorieTargetKcal": {
            "type": "integer",
            "minimum": 1000
          },
          "estimatedExpenditureKcal": {
            "type": "integer",
            "minimum": 1
          },
          "estimatedWeeks": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "rateLimited": {
            "type": "boolean"
          },
          "restingEnergyKcal": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "birthDate",
          "currentWeightKg",
          "dailyActivity",
          "exerciseFrequency",
          "heightCm",
          "metabolicSex",
          "targetWeightKg",
          "weeklyWeightChangePercent",
          "weightGoal",
          "calculatedAt",
          "calorieTargetKcal",
          "estimatedExpenditureKcal",
          "estimatedWeeks",
          "rateLimited",
          "restingEnergyKcal"
        ]
      },
      "NutritionPlanResponse": {
        "type": "object",
        "properties": {
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NutritionPlan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "plan"
        ]
      },
      "SaveNutritionPlan": {
        "type": "object",
        "properties": {
          "birthDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "currentWeightKg": {
            "type": "number",
            "minimum": 30,
            "maximum": 350
          },
          "dailyActivity": {
            "enum": [
              "mostly_sedentary",
              "moderately_active",
              "very_active"
            ],
            "type": "string"
          },
          "exerciseFrequency": {
            "enum": [
              "none",
              "one_to_three",
              "four_to_six",
              "seven_plus"
            ],
            "type": "string"
          },
          "heightCm": {
            "type": "number",
            "minimum": 100,
            "maximum": 250
          },
          "metabolicSex": {
            "enum": [
              "female",
              "male"
            ],
            "type": "string"
          },
          "targetWeightKg": {
            "type": "number",
            "minimum": 30,
            "maximum": 350
          },
          "weeklyWeightChangePercent": {
            "type": "number",
            "minimum": 0,
            "maximum": 1.25
          },
          "weightGoal": {
            "enum": [
              "lose",
              "maintain",
              "gain"
            ],
            "type": "string"
          }
        },
        "required": [
          "birthDate",
          "currentWeightKg",
          "dailyActivity",
          "exerciseFrequency",
          "heightCm",
          "metabolicSex",
          "targetWeightKg",
          "weeklyWeightChangePercent",
          "weightGoal"
        ]
      },
...
      "FoodLogResponse": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FoodLogEntry"
            }
          }
        },
        "required": [
          "entries"
        ]
      },
...
    }
  }
}